// src/types.ts
export interface IProduct {
  id: string;
  description?: string;
  image?: string;
  title: string;
  category?: string;
  price: number | null;
}

export interface IOrderForm {
  payment?: string;
  address?: string;
  email?: string;
  phone?: string;
  total?: number | string;
}

export interface IOrder extends IOrderForm{
  items: string[];
}

export interface IAppState {
  catalog: IProduct[];
  basketList: IProduct[];
  preview: string | null;
  order: IOrder | null;
  loading: boolean;
}

export interface IApi {
  baseUrl: string;
  requestOptions: string;
}